# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API of Alexandria.",
        "title": "Alexandria API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v2",
    "paths": {
...
        "/versions": {
            "post": {
                "description": "Create a new version with discussions and repository from zipped file in body",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "versions"
                ],
                "summary": "Create new version",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Repository to create",
                        "name": "repository",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.VersionDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/versions/{versionID}": {
            "get": {
                "description": "Get a version by version ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "versions"
                ],
                "summary": "Get version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Version ID",
                        "name": "versionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.VersionDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/versions/{versionID}/discussions": {
            "get": {
                "description": "Returns all discussions on this version that are not a reply to another discussion\nEndpoint is offset-paginated",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "versions"
                ],
                "summary": "Returns all level 1 discussions associated with the version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "version ID",
                        "name": "versionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page query",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DiscussionDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/versions/{versionID}/file/{filepath}": {
            "get": {
                "description": "Get the contents of a single file from a repository of a version",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "versions"
                ],
                "summary": "Get a file from a repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Version ID",
                        "name": "versionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filepath",
                        "name": "filepath",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/versions/{versionID}/render": {
            "get": {
                "description": "Get the render of the repository underlying a version if it exists and has been rendered successfully",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "versions"
                ],
                "summary": "Get the render of a version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Version ID",
                        "name": "versionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/versions/{versionID}/repository": {
            "get": {
                "description": "Get the entire zipped repository of a version",
                "produces": [
                    "application/zip"
                ],
                "tags": [
                    "versions"
                ],
                "summary": "Get the repository of a version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Version ID",
                        "name": "versionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/versions/{versionID}/tree": {
            "get": {
                "description": "Get the file tree of a repository of a version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "versions"
                ],
                "summary": "Get the file tree of a repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "version ID",
                        "name": "versionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.DiscussionDTO": {
            "type": "object",
            "properties": {
                "anonymous": {
                    "type": "boolean"
                },
                "deleted": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "memberID": {
                    "type": "integer"
                },
                "replyIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "text": {
                    "type": "string"
                },
                "versionID": {
                    "type": "integer"
                }
            }
        },
...
        "models.RenderStatus": {
            "type": "string",
            "enum": [
                "success",
                "pending",
                "failure"
            ],
            "x-enum-varnames": [
                "RenderSuccess",
                "RenderPending",
                "RenderFailure"
            ]
        },
...
        "models.VersionDTO": {
            "type": "object",
            "properties": {
                "discussionIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "renderStatus": {
                    "$ref": "#/definitions/models.RenderStatus"
                }
            }
        },
...
        "utils.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}